+++ packages/core/src/types.ts
@@ -64,6 +64,8 @@ export type Execution = ExecaReturns;
 
 export type ExecuteType = 'parallel' | 'pool' | 'serial' | 'sync';
 
+export type StdioType = 'buffer' | 'stream' | 'inherit';
+
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace NodeJS {